cmake: do not unconditionally find OpenGL package
The OpenGL package provides libGL and (Desktop) OpenGL header files.
libGL actually provides two interfaces:
  GLX "window system layer" (as defined by glproto)
  Desktop GL "GL Context API"

For environments that do not use either GLX nor Desktop GL, there is no
reason to look for this package.  In fact, if OpenGL is not found, the
following cmake variables get set to "NOTFOUND":
  OPENGL_INCLUDE_DIR
  OPENGL_gl_LIBRARY

OPENGL_INCLUDE_DIR as NOTFOUND is a problem when building EGL tests and
the piglitutil_gles* libraries, since OPENGL_INCLUDE_DIR is the common
UTIL_GL_INCLUDES, which is in the list of included directories.  It would
be better if it just wasn't defined at all.

Similarly, OPENGL_gl_LIBRARY is linked to the egl_khr_create_context tests
to enable them to use EGL to create a Desktop GL context.  Again, if there
is no Desktop GL, it would probably be better to just not define it at all.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>