core: fix remove_tests logic
The sole user is tests/gpu.py, which was passing in a predicate that
worked (a) as a filter, and (b) expected to receive a test. However the
implementation operated on the dict-of-dicts. Instead keep track of all
the filtering functions, and apply the filters when flattening the tests
tree.

Take this opportunity to pass in a second argument to the filter
function: the full test path. This should allow future conversion of the
'del' style of filtering.

Signed-off-by: Ilia Mirkin <imirkin@alum.mit.edu>
Reviewed-by: Dylan Baker <baker.dylan.c@gmail.com>