framework/summary.py: add assert to Summary.__init__()
Normally piglit doesn't generate test names with leading slashes, and
the code has no way to deal with them. In fact, a leading slash will
cause a loop in Summary.__init__ to run infinitely.  The simplest
solution is an assertion, which means the program will assert before it
ever gets to the loop.

Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>