layered-rendering/blit: specify the final fb grid using tex dimensions
This made it as least for me easier to read the individual functions
in isolation without the need of jumping back and forth checking
the definitions of "half dimensions" which in turn got divided by two
from time to time.

The sizes of the textures couldn't be changed anyway without altering
these window dimensions based arguments correspondingly.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>
Signed-off-by: Topi Pohjolainen <topi.pohjolainen@intel.com>