framework/core: Fix performance regression from e4c9a3621a
There is a sort of corner case regression caused by this patch, if the
number of tests are less than 50 they are effectively always run
serially. This can have a pretty big performance impact, especially if
any of those tests takes a long time to run. Reducing this to 1 results
in a very slight performance reduction of very long runs (~2 seconds
difference on quick.tests), but the trade off for short runs is much
more significant.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>