max-texture-size: Convert to using subtests.
This test currently checks 30 different subcases.  Prior to this patch,
if one of them failed, the whole test would fail.  Running it and
examining the output would tell you which subcases had trouble, but
there was nothing in the actual Piglit summary.

Currently, this reports 'fail' for a legitimate fail, and 'pass'
otherwise.  The next commits will refine this.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Matt Turner <mattst88@gmail.com>