max-texture-size: Report SKIP and stop on GL_OUT_OF_MEMORY errors.
Previously, the test would call TexImage to create the texture image.
If it received a GL_OUT_OF_MEMORY error, it would continue on and call
TexSubImage to try and overwrite a part of the image...which doesn't
exist, because the GL ran out of memory when trying to create it.

Marek tried to work around this by making it accept INVALID_VALUE for
the TexSubImage calls, but Mesa can also report INVALID_OPERATION in
some cases.  Rather than quibbling about error codes, we should step
back and take a sensible action: if a texture image doesn't exist,
don't try and write to it.

When we receive a GL_OUT_OF_MEMORY error, we now report a subtest
result of "skip" and move on.  We can't realistically claim "pass" or
"fail", since we have no idea whether the implementation works or not.

For other errors, we report "fail" and stop.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Matt Turner <mattst88@gmail.com>