glx_oml_sync_control: report FAIL if glXGetMscRateOML returns an error
The glx_oml_sync_control does not parse parameters.  In particular, it
does not parse the "-auto" flag to set piglit_automatic.

So, when we return FAIL from draw(), it returns to the
piglit_glx_event_loop, which just continues to spin waiting for events.
Instead, just report the FAIL result, which quits the program.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>