util/egl: Add EGLDisplay parameter to piglit_require_egl_extension()
Without an explicit EGLDisplay parameter, it's not possible to call
piglit_require_egl_extension() without first binding the display to
a context with eglMakeCurrent(). Some EGL tests, though, will never need
to or want to create a context.

Also, this makes the signature of piglit_require_egl_extension()
match that of piglit_is_egl_extension_supported().

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>