texturing/s3tc-errors: double check GL_TEXTURE parameters and alloc
If the GL does not actually support one of the compression formats in
s3tc_formats, it may silently choose a different actual internalformat.

In such a case, the resulting compressed image size may be larger,
leading to memory corruption when the bigger-than-expected image is read
back with glGetCompressedTexImage() into a buffer that was allocated to
the expected, smaller, size.

Instead, first confirm the texture has really been compressed, that the
format is as expected, and that the size matches our pre-computed
expected size.  If any of these fail, set pass to false and spit an
error, but keep going, using the actual format and size returned by the
GL.

Ian Romanick pointed out that "if we're going to all this effort, we
should probably go all the way. A follow-on patch should check that
glGetCompressedTexImage doesn't over-run the buffer."

v2:
 * Fixed parameter check block format per Ian Romanick review
 * Updated commit message title per Brian Paul review
 * Function call continuation lines align with open '(' (like in rest of
   file).

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>