python: Convert from print statement to function
One of the biggest changes between python2 and python3 is that print
changed from a statement to a function. This change was backported to
python2 in the __future__ module, and this patch makes that conversation
to ease the transition from python2 to python3.

v2: - Replace dict comprehension with dict(generator). This is for
      python 2.6 compatability

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Ian Romanick <ian.d.romanick@intel.com> (v1)