gl-2.0: Verify results of program validation when a conflicting sampler configuration is used.
Section 2.15.4 (Shader Execution), subheading "Validation" of the OpenGL
2.0 spec says:

    "[INVALID_OPERATION] is generated by Begin, RasterPos, or any
    command that performs an implicit Begin if:

        - any two active samplers in the current program object are of
          different types, but refer to the same texture image unit,"

This test verifies this behavior several ways.  First, an invalid
configuration is constructed.  When the program is in the invalid
configuration, glValidateProgram is used to determine the status.

The program is then transitioned to a valid configuration, and
glValidateProgram is called again.

The program is then transitioned back to the invalid configuration.
Without calling glValidateProgram, glBegin is called.  This should
generate the aforementioned error.  While still in the invalid state
glValidateProgram is called a third time.

Finally, the program is transitioned back to the valid configuration.
Without calling glValidateProgram, glBegin is called.  This should not
generate any error.

All of the state flip-flopping is done in an attempt to catch
implementations that latch state in glValidateProgram and do not
re-validate in glBegin.

NOTE: Mesa fails this test because it does not generate
GL_INVALID_OPERATION the first time glBegin is called.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Eric Anholt <eric@anholt.net>
Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>