sso: Verify results of program pipeline validation when a conflicting sampler configuration is used.
Section 2.11.11 (Shader Execution), subheading "Validation," of the
OpenGL 4.1 spec says:

    "[INVALID_OPERATION] is generated by any command that transfers
    vertices to the GL if:

        ...

        - Any two active samplers in the current program object are of
          different types, but refer to the same texture image unit."

This test verifies this behavior several ways.  First, an invalid
configuration is constructed.  When the pipeline is in the invalid
configuration, glValidateProgramPipeline is used to determine the
status.

The pipeline is then transitioned to a valid configuration, and
glValidateProgramPipeline is called again.

The pipeline is then transitioned back to the invalid configuration.
Without calling glValidateProgramPipeline, glDrawArrays is called.  This
should generate the aforementioned error.  While still in the invalid
state glValidateProgramPipeline is called a third time.

Finally, the pipeline is transitioned back to the valid configuration.
Without calling glValidateProgramPipeline, glDrawArrays is called.  This
should not generate any error.

All of the state flip-flopping is done in an attempt to catch
implementations that latch state in glValidateProgramPipeline and do not
re-validate in glDrawArrays.

NOTE: The work-in-progress GL_ARB_separate_shader_objects implementation
for Mesa currently fails this test because glValidateProgramPipeline
always succeeds, and it does not generate GL_INVALID_OPERATION the first
time glDrawArrays is called.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Eric Anholt <eric@anholt.net>
Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>