util/egl: Add piglit_egl_get_default_display()
This function gets the default display for the given platform.

If the given platform platform is EGL_NONE, the the function wraps
eglGetDisplay().  Otherwise, it wraps eglGetPlatformDisplayEXT().  If
EGL does not support the platform extension for the given platform, then
it returns EGL_NO_DISPLAY.

This is useful for EGL tests that need to create a display for
a specific platform.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>