glsl_parser_test.py: Simplify GLSLParserTest
This vastly simplifies the way that GLSLParserTest works. This
simplification has a couple of advantages. First, we call
PlainExecTest's __init__ rather than ExecTest's __init__, which is good
OO practice; second we reduce the number of regex's used, which yields
enhanced performance. In tests this reduced GLSLParserTest
initialization time by about 40%

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>