ext_timer_query/time-elapsed: Prevent infinite loop.
When testing with llvmpipe, somehow the draw() function is returning
constant elapsed time of 0.000222 approximately, regardless how big
iters is.

This is probably a bug in llvmpipe, but it seems a good idea to fix the
test to prevent integer overflow and the consequent infinite loop in the
test regardless.

Reviewed-by: Brian Paul <brianp@vmware.com>