framework/exectest.py: remove deprecated builtin use
Upstream python considers map() and filter() deprecated in favor of
using comprehensions. These comprehensions are easier to read, and can
act as both a map and a filter at the same time. In this case we are
using generator comprehensions, which have an additional advantage of
lower memory consumption.

v2: - minor readability improvements

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>