framework/exectest.py: fix check_for_skip_scenario
This makes a couple of changes. First it removes Test.skip_test, which
was set equal to Test.check_for_skip_scenario(). Second, it removes the
unused command argument from check_for_skip_scenario's signature.
Finally it implements a no-op version in Test, and moves the version in
Test to PiglitTest, since it is a native piglit test specific check.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>