framework: Replace Group() with dict()
Group is a subclass of dict, with no overrides. This is silly and
confusing, it's a dictionary, just use a dictionary. This has the
additional advantage of removing a lot of spaghetti dependency issues,
since nearly every module needed access to Group()

Most of this was done with the shell command:
find . -name '*py' | xargs sed -i -e 's!Group()!{}!g'

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>