framework/exectest.py: Replace if with assert in Test
It is more useful to put a check in the Test.command property that the
internal name (Test._command) is not empty than that the returned value
(which is altered by Test.command) is not None.

This changes the behavior in that now the python will raise an uncaught
exception rather marking the test as skip.

v2: - Replace a similar patch that was later in the series with this

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>