framework: small refactor of test_result page generation
Previously each value of a dictionary was pulled out and passed one by
one to mako, that results in a lot of values being passed. With this
patch the dictionary is passed and the keys are pulled out in the mako.
This is much cleaner and easier to read.

v2: - fix dmesg is list problem by altering the dictionary

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>