profile.py: Cleanup a few pylint warnings
This adds self._dmesg definition to the TestProfile constructor (this
isn't necessary but it's good form), and renames self.__dmesg to
self._dmesg. A double underscore is not about making an name or method
private, it's about making it invisible to that class's children. We
don't want to hide self._dmesg from any children of TestProfile.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>