piglit: Add a toplevel piglit command
This uses the functions created by the last commit to create a
all-in-one piglit command. It uses argparse to parse partial commands
and provide help for those functions.

It is also capable of handling all of the black magic for setting the
python sys.path and setting the PIGLIT_SOURCE_DIR environment variable.

v2: - Allow the name of the program to not be piglit (example:
      piglit-20140101)
    - Remove unused check from cmake
    - Install piglit as piglit${PIGLIT_INSTALL_VERSION}; when this
      variable is unset piglit will be installed simply as 'piglit'
v3: - split "../lib" into "..", "lib" in os.path.join()

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Jordan Justen <jordan.l.justen@intel.com>