outerproduct_tests: generate tests
This replaces outerProduct.sh and outerProduct-const.sh with a single
python generator run at build time. This removes a significant amount of
duplicate code that is otherwise checked into git, combines two
generators that are largely the same, and being written in mako is
easier to read and edit.

v2: - fix some typos in a generated comment
    - preserve newline between probe and next uniform
    - fix a few other minor whitespace problems

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Acked-by: Kenneth Graunke <kenneth@whitecape.org>