framework/exectest.py: Fix a few pylint warnings
This removes some unreachable code, adds an name definition to the Test
constructor, and comments a bare except. There are only a couple of
cases where a bare except is appropriate, most of the time it is a
problem, but we are using one the few cases where it is okay, we're
logging all exceptions to a file.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>