gleantest.py: simplify interpret_result()
First, this replaces str.find() with 'if x in str', this is simpler and
easier to read, and unlike find(), short-circuits on the first True.
Second, it removes the check for skip. This isn't needed since if the
test is skipped it will return before reaching the interpret_result()
call, making this dead code.

v2: - split this change out of the next patch

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>