builtin-gl-sample-mask-simple: another test for gl_SampleMask[0] FS output
This one is simpler and easier to debug. It passes on my hardware, while
the other one fails (due to a shader bug probably).

I don't like tests that return the pass or fail status from the shader,
because they are impossible to debug. Therefore, this one doesn't do it.

Reviewed-by: Chris Forbes <chrisf@ijw.co.nz>