Test GL_ARB_clear_texture using cube map textures
This adds a test for glClearTexSubImage when clearing a sub-region of
a cube map texture. Each face of the cube map is cleared to a separate
colour and then one pixel of each face is rendered using points. The
colours are then probed to check whether the clear worked. This is
important to test because glClearTexSubImage has some unusual
semantics in that the zoffset parameter is used to select the cube
face.

Reviewed-by: Brian Paul <brianp@vmware.com>