Test GL_ARB_clear_texture using a range of base formats
This adds a test for glClearTexSubImage when clearing a sub-region of
a 2D texture using a range of formats that are supported in basic GL.
A utility function is added to test a format combination so that
further tests can be added with other formats that need an extension.

For each format a texture is created and two sub-regions are cleared.
One region is set to all zeroes using NULL as the clear value and the
other is set to a known value. The texture is then retrieved with
glGetTexImage and compared against the expected values.

Reviewed-by: Brian Paul <brianp@vmware.com>