Test GL_ARB_clear_texture using 3D textures
This adds a test for glClearTexSubImage when clearing a sub-region of
a 3D texture. Two sub-regions are cleared, one using a NULL data
parameter and one providing a colour. Both regions have a depth
greater than 1. All of the images of the texture are then rendered to
the window and probed for the correct results.

Reviewed-by: Brian Paul <brianp@vmware.com>