gleantest_tests.py: Make one test more robust
The test for glean returncodes relied on unsetting os.environ to cause
glean to fail. This is fairly reliable, but it is possible to break it.
A much more robust implementation is to just set
instance.result['returncode'] = 1 and then run interpret_result(), with
the added bonus of focusing the test better.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>