utils.py: add boilerplate to work around nose bug
The gist of the bug is that the way nose uses generators to create tests
it yields the same instance multiple times, and it uses an instance
attribute to set the name of the test. It then prints the results of the
tests all at once, meaning that the final assignment of that instance
attribute is displayed for all failing tests.

The class is a proxy object that is used to create a unique instance for
each test. This works around the bug since the attribute that stores the
test description isn't overwritten each time a new test is generated.

The decorator is used for generators to transparently replace the
function yielded by the test generator with a proxy object. This allows
generated tests to return the correct name; all without any modification
to the test generator. This means that if nose ever solves this problem
making use of nose's solution is a single deleted line per generator.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>