dmesg.py: Use Abstract Base Class
This change isn't really a functional change, but it should allow
someone to implement a dmesg class for another OS (OSX, BSD, Linux
without timestamps, etc) more readily, since it simplifies subclassing.

The immediate reason for doing this however, is that it creates a clean
separation between update_dmesg() and update_result(), which makes it
easier to test update_result() without needing privileged access (or
even a posix machine).

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>