exectest.py: Return early if _run_command() hits the except block
If the except block is hit in Test._run_command(), the test has a status
of skip, but the test continues to run anyway. This is not consistent
with the way Test.check_for_skip_scenario() works (which returns early
if true). The effect of this inconsistency is that tests with a known
value still call interpret_result(). This wastes time in python
processing.

This patch causes a failure in Test._run_command() to set
self.result['result'] to 'skip' (it should be 'fail' before this). When
Test._run_command() is called by Test.run(), if self.result['result'] is
'skip' then run() will return early.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>