Test Reporting: Use JSON as interchange format for piglit tests
It just so happens that the plain text representation of a python
dictionary is a superset of a JSON dictionary. The only real difference
is that python allows ' and " to be used interchangeably while JSON only
accepts ". It also just so happens that JSON is a really nice format for
sending data between two programs, and that python has a very fast
implementation of JSON available, which piglit already supports
(simplejson) as well as a native (but slower) version.

This patch removes the use of eval(), which is considered slow, and
allows for arbitrary code execution. By moving to JSON as our
interchange format between compiled tests and python we gain
standardization, easier maintenance, and less python code required.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>