Test Reporting: simplify subtest reporting
Since we're using json to decode our results, we can simplify the way we
pass information between C/C++ tests and the python framework. This
replaces the PIGLIT:subtest {"<name>": "<status>"} syntax with
PIGLIT: {"subtest": {"<name>": "<status>"}}.

This is a very clean solution, and eliminates the need for several if
branches and loop iterations in the python framework.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>