GL_ARB_texture_multisample: Test using glClear with a ms render target
This creates a mulitsample texture and tries clearing it with glClear
via a framebuffer. One of the clears covers the entire texture and the
other is clipped to the right half of the texture using a scissor.

On the i965 driver the clear that covers the whole of the texture
wasn't working in Mesa before commit 765efeef883cdbb86e052acd887e0c3.

https://bugs.freedesktop.org/show_bug.cgi?id=79729