framework/summary/html: Prevent `\` in URLs.
It's fine for URLs to have `\` directory separators when viewing the
HTML files on Windows.  But if one moves the HTML report to a Linux
machine or puts it on a web-server and accesses it from a Linux machine
then links with `\` in them will become dead.

This change prevents that by always replacing `\` with `/` in URLs.

(There might be an easier way to achieve this than to have so many
instances of `foo.replace('\\', '/')`.  For example, it might be easier
to do this replacement when loading the results.  On the other hand,
there are several places that use os.path, so it might not solve the
issue completely neither.)

v2: Avoid wrong backward/forward terminology in comment, as pointed out
by Brian.