piglit (executable): Make PIGLIT_SOURCE_DIR detection more robust
The initial implementation I committed was a bit hackish and I knew it
at the time, but it seemed to work for all the cases that I tested.
However, there were corner cases. The biggest of these cases is problems
running out of tree with piglit built in tree.
ex: src/piglit/piglit run ...

This patch makes the handling more robust by explicitly searching for
three distinct cases:
1) piglit is run from the source dir, built in the source dir
2) piglit is run from outside the source dir, built in the source dir
3) piglit has been installed out of tree

I have tested all three of these cases, and case 3 if piglit is
installed as piglit.py.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Jose Fonseca <jfonseca@vmware.com>