fbo: test that swizzle doesn't affect mipmap generation
This is a regression test for a mesa meta bug fix whereby a swizzle set
via GL_ARB_texture_swizzle was being applied while iteratively
generating a mipmap. This meant the swizzle would be applied twice by
the time the application sampled those generated layers.

The bug was originally noticed with code that used a swizzle to emulate
a component alpha texture via a component red texture. In this case the
texture internally only had one component and applying the swizzle
effectively discarded that component. The test follows a similar
pattern, but swizzles from red to green instead of to alpha.

Reviewed-by: Brian Paul <brianp@vmware.com>