framework: Don't rely on os.environ so much
There are some problems with using os.environ, all of which are caused
shell child-variable relationships. Most of these problems are easy to
solve, since we don't actually need to have environment variables set
except during test execution, and that can be passed to Popen (or it's
helper wrappers). This gives us much better assurance that things are
happening in the way we expect.

v7: - use itertools.chain instead of nested iterations
    - fix piglit.conf default case
    - reorder environment overrides so that variables passed in by the
      user (or exported) override the builtin values instead of
      vice-versa

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>