core: replace Environment.collectData()
Environment.collectData() is strange and awful for a host of reasons.
1) Its implementation was such that it should have been a static method,
   but it wasn't
2) It wasn't at all related to Environment
3) It used a public helper which it was the sole consumer of.

Replacing it with a top level function makes a lot more sense for all of
these reason. This function has another major advantage, it assumes
nothing. In the former implementation it was assumed that specific
operating systems had specific tools available, and that those tools
were only available on that platform. This implementation doesn't assume
that, instead it makes a blind leap that the tools is available, and
then is caught if its wrong.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>