cl: Add signbit(float) relational tests
Also, while doing this, needed to add a NEGNAN define to genclbuiltins.py because
python just prints nan for float("-nan") when generating tests.

cl-program-tester can handle -nan as an input correctly, so just pass the value
through as a string without python attempting to handle it.

v2: Add nan and -nan tests

Signed-off-by: Aaron Watry <awatry@gmail.com>
Reviewed-by: Tom Stellard <thomas.stellard@amd.com>
CC: Tom Stellard <thomas.stellard@amd.com>