arb_draw_indirect: Verify gl_VertexID when used with indirect rendering
When rendering with glDrawArraysIndirect, the value of gl_VertexID
observed in the shader should start with the value of 'first' and
increment from there.

When rendering with glDrawElementsIndirect, the value of gl_VertexID
observed in the shader should be the value retrieved from the index
buffer plus the value of basevertex.

Passes on AMD R9 270X (Pitcairn) with Catalyst 14.6 beta 1.0-jul11.

NOTE: This test currently fails on the i965 driver.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Tested-by: Kenneth Graunke <kenneth@whitecape.org>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=80247