glx-swap-event: Make sure that swap events are on the correct drawable
DRI2 had a bug in its implementation which caused the
GLX_BufferSwapComplete event to be emitted for the X11 Drawable,
instead of the GLXDrawable instead, as the specification for
INTEL_swap_event says should happen.

This didn't happen for indirect GLX contexts or for DRI3, which properly
sent out the GLXDrawable inside the code. Since userspace already relies
on getting the X11 Drawable, we are fixing the spec and mesa to be based
around the X11 Drawable instead.

Make sure that the event has the X11 Drawable in its drawable field.

Corresponds to mesa commit b4dcf87f34f68111acd9d364739938721fc692f4

Signed-off-by: Jasper St. Pierre <jstpierre@mecheye.net>
Signed-off-by: Dave Airlie <airlied@redhat.com>