results.py: Fix JSONWriter.close_json()
A previous patch created a bug which caused an AssertionError at the end
of a piglit run. This bug is the result of calling
JSONWriter.close_dict() too many times. This solution is to split the
start of the test dict out of initialize_json() and do it in the calling
function instead.

v2: - take this approach rather than close dicts as a stack

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>
Tested-by: Tom Stellard <thomas.stellard@amd.com>