piglit-dispatch: fix get_core_proc_address() for Windows
Even for functions such as glGetStringi() which are in OpenGL 3.0, we
call get_core_proc_address() with gl_10x_version = 10 (seems to be a
larger problem in the dispatch code generation).  That means we try
GetProcAddress("OPENGL32.DLL") and fail.  Now, if GetProcAddress()
fails, try get_ext_proc_address() too.

Reviewed-by: Charmaine Lee <charmainel@vmware.com>