amd_vertex_shader_viewport_index: add render test
This is a copy of the render_viewport ARB_viewport_array test. I didn't
think that it was worth it to copy the scissor/depth tests, since they
basically test the same thing but making sure that scissor/depth are
applied correctly. If that works for ARB_viewport_array, it really
should work here too.

Signed-off-by: Ilia Mirkin <imirkin@alum.mit.edu>
Reviewed-by: Chris Forbes <chrisf@ijw.co.nz>