util/gl: Unify piglit_checkerboard_texture between GL and GLES
In the GLES implementation, factor out the conversion of the white and
black pixel values from float to ubyte outside the loop.  That then
unifies the loop between the two implementations.

Update the documentation to note that piglit_checkerboard_texture sets
GL_CLAMP_TO_EDGE on GLES, not GL_CLAMP_TO_BORDER as it does on GL.

Signed-off-by: Josh Triplett <josh@joshtriplett.org>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>