glsl_parser_test.py: Don't allow duplicate keys
If a single key is provided more than once raise an exception. This
behavior is desirable because the implementation (both before and
after this series), simply takes the last value for a given key. So,
providing two copies of the same key will result in the last one being
honored and the first being ignored.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>